Move new manifest functions to source/addonStore/models/manifest.py
